% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochasticVolatility_forward.R
\name{simulateForwards}
\alias{simulateForwards}
\title{Simulate a stochastic volatitlity modell forwards}
\usage{
simulateForwards(steps = steps_D, x0 = x0_D, B = B_D, U = U_D,
  psi = psi_D, phi = phi_D)
}
\arguments{
\item{steps}{Number of timesteps to be taken}

\item{x0}{initial conditions}

\item{B}{load matrix.}

\item{U}{covariance matrix of x-update}

\item{psi}{weights added to covariance-matrix of observations}

\item{phi}{x-rescaling before updating}
}
\value{
Returns a list of two matrices "X" and "Y". The column-index of each matrix corresponds to time.
}
\description{
Simulates A stochastic volatilityu model as outlined in section 4.2 of (Lee et al. 2010).

x[t] ~ Normal(diag(phi)*x[t-1], U)

y[t] ~ Normal(0, B * diag(exp(x[t])) * B' + diag(psi)).

All arguments default to the arguments used in (Lee et al. 2010).
}
\author{
Mathias C. Cronj\"ager
}

